// wavue/packages/wagrid/src/core/columns/WaGridColumn.enum.ts
export enum WaColumnKind {
    column = 'column',
    header = 'header',
    empty  = 'empty',
}
export enum WaColumnProperty {
    rowId = '_rowId',
    rowMode = '_mode',
    isChecked = '_isChecked',
    num = '_number',
    mode = '_mode',
    checkbox = '_checkbox',
    parentNum = '_parentNumber',
    depth = '_depth',
    children = 'children',
    childRowIds = '_childRowIds',
    childRows = '_childRows',
    childCount = '_childCount',
    isOpen = '_isOpen',
    isShow = '_isShow',
    open = 'open',
    closed = 'closed',
    rowCount = '_rowCount',

    /** User Property */
    name = 'name',
    text = 'text',
    type = 'type',
    dataType = 'dataType',
    width = 'width',
    editable = 'editable',
    visible = 'visible',
    align = 'align',
    scale = 'scale',
    roundType = 'roundType',
    fixedScale = 'fixedScale',
    scaleMax = 'scaleMax',
    scaleMin = 'scalemin',
    showZero = 'showZero',
    commaUnit = 'commaUnit',  // to be deprecated, Fixed 3
    thousandChar = 'thousandChar',
    decimalChar = 'decimalChar',
    currencyChar = 'currencyChar',
    className = 'className', // className
    resizable = 'resizable',
    sortable = 'sortable',
    movable = 'movable',
    autoResizable = 'autoResizable',
    autoWidth = 'autoWidth',
    summaryType = 'summaryType',
    required = 'required',
    combo = 'combo',
    format = 'format',
    kind = 'kind', //header, column, empty
    rowSpan = 'rowSpan',
    colSpan = 'colSpan',
    rowIndex = 'rowIndex',
    colIndex = 'colIndex',
    subRowSpan = 'subRowSpan',
    subColSpan = 'subColSpan',
    order = 'order',
    value = 'value',
    header = 'header',
}